+++ src/app/app.module.ts
@@ -8,6 +8,7 @@ import { MatSnackBarModule } from '@angular/material/snack-bar';
 import { MatToolbarModule } from '@angular/material/toolbar';
 import { MatCardModule } from '@angular/material/card';
 import { MatChipsModule } from '@angular/material/chips';
+import { EverythingLibModule } from '@zwiqler94/everything-lib';
 
 import { MatTabsModule } from '@angular/material/tabs';
 import { MatListModule } from '@angular/material/list';
@@ -26,6 +27,9 @@ import { getStorage, provideStorage } from '@angular/fire/storage';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatInputModule } from '@angular/material/input';
+import { MatSidenavModule } from '@angular/material/sidenav';
+import { MatExpansionModule } from '@angular/material/expansion';
+import { MatIconModule } from '@angular/material/icon';
 
 import { AppComponent } from './app.component';
 import { AboutMeComponent } from './pages/about-me/about-me.component';
@@ -41,6 +45,9 @@ import { CarouselComponent } from './components/carousel/carousel.component';
 import { environment } from 'src/environments/environment';
 import { ServiceWorkerModule } from '@angular/service-worker';
 import { LoginPageComponent } from 'src/app/pages/login-page/login-page.component';
+import { PostBaseComponent } from 'src/app/components/post-base/post-base.component';
+import { ContactMeComponent } from 'src/app/pages/contact-me/contact-me/contact-me.component';
+import { FooterComponent } from './components/footer/footer.component';
 
 declare global {
   // eslint-disable-next-line no-var
@@ -61,8 +68,11 @@ self.FIREBASE_APPCHECK_DEBUG_TOKEN = environment.appCheckDebug;
     MainFeedPageComponent,
     TextPostComponent,
     LinkPostComponent,
+    PostBaseComponent,
     CarouselComponent,
     LoginPageComponent,
+    ContactMeComponent,
+    FooterComponent,
   ],
   imports: [
     BrowserModule,
@@ -80,6 +90,9 @@ self.FIREBASE_APPCHECK_DEBUG_TOKEN = environment.appCheckDebug;
     MatChipsModule,
     FormsModule,
     ReactiveFormsModule,
+    MatSidenavModule,
+    MatExpansionModule,
+    MatIconModule,
     provideFirebaseApp(() => initializeApp(environment.firebaseConfig)),
     provideAppCheck(() =>
       initializeAppCheck(getApp(), {
@@ -92,7 +105,6 @@ self.FIREBASE_APPCHECK_DEBUG_TOKEN = environment.appCheckDebug;
     provideAnalytics(() => initializeAnalytics(getApp())),
     provideAuth(() => {
       const auth = getAuth();
-
       if (environment.local) {
         connectAuthEmulator(auth, 'http://localhost:9099', {
           disableWarnings: true,
@@ -101,11 +113,12 @@ self.FIREBASE_APPCHECK_DEBUG_TOKEN = environment.appCheckDebug;
       return auth;
     }),
     ServiceWorkerModule.register('ngsw-worker.js', {
-      enabled: environment.production,
+      enabled: environment.serviceOptions.useServiceWorker,
       // Register the ServiceWorker as soon as the application is stable
       // or after 30 seconds (whichever comes first).
-      registrationStrategy: 'registerWhenStable:30000',
+      registrationStrategy: 'registerWhenStable:20000',
     }),
+    EverythingLibModule,
   ],
   providers: [],
   bootstrap: [AppComponent],