+++ src/app/components/link-post/link-post.component.ts
@@ -1,5 +1,6 @@
 import { Component, Input, OnInit } from '@angular/core';
 import { LinkPreview } from 'src/app/components/models/link-post';
+import { PostBaseComponent } from 'src/app/components/post-base/post-base.component';
 import { LinkPreviewService } from 'src/app/services/link-preview/link-preview.service';
 
 @Component({
@@ -7,12 +8,12 @@ import { LinkPreviewService } from 'src/app/services/link-preview/link-preview.s
   templateUrl: './link-post.component.html',
   styleUrls: ['./link-post.component.scss'],
 })
-export class LinkPostComponent implements OnInit {
-  @Input() postTitle: string;
-  @Input() postContent: string;
+export class LinkPostComponent extends PostBaseComponent implements OnInit {
   linkPreviewData: LinkPreview;
 
-  constructor(private linkPreviewService: LinkPreviewService) {}
+  constructor(private linkPreviewService: LinkPreviewService) {
+    super();
+  }
 
   get getLinkPreviewData() {
     return this.linkPreviewData;
@@ -23,21 +24,34 @@ export class LinkPostComponent implements OnInit {
   }
 
   async ngOnInit() {
-    await this.getLinkPreview();
+    try {
+      await this.getLinkPreview();
+    } catch (err) {
+      console.error(err);
+    }
   }
 
   async getLinkPreview() {
     const linkArray = this.postContent.match(
-      /(http|https):\/\/www.[a-z]*.[a-z]*\/[a-zA-Z0-9?=]*/,
+      /(http|https):\/\/(www\.)?[a-zA-Z0-9]+\.[a-zA-Z0-9]+[a-zA-Z0-9/\-.,&?=%#();:~]*/,
     );
     if (linkArray !== null) {
-      (
-        await this.linkPreviewService.getLinkPreview(String(linkArray[0]))
-      ).subscribe((data: unknown) => {
-        {
-          this.linkPreviewData = data as LinkPreview;
-        }
-      });
+      try {
+        (
+          await this.linkPreviewService.getLinkPreview(String(linkArray[0]))
+        ).subscribe({
+          next: (data: unknown) => {
+            {
+              this.linkPreviewData = data as LinkPreview;
+            }
+          },
+          error: (err) => {
+            throw new Error(JSON.stringify(err));
+          },
+        });
+      } catch (err: any) {
+        throw new Error(err);
+      }
     }
   }
 }