+++ src/app/pages/about-me/about-me.component.ts
@@ -1,29 +1,69 @@
 /* eslint-disable @typescript-eslint/naming-convention */
 /* eslint-disable no-underscore-dangle */
 import { HttpClient } from '@angular/common/http';
-import { Component, VERSION } from '@angular/core';
+import { Component } from '@angular/core';
 import { FormBuilder, FormGroup } from '@angular/forms';
-
+import { MatTabChangeEvent } from '@angular/material/tabs';
+import { environment } from 'src/environments/environment';
 @Component({
   selector: 'app-about-me',
   templateUrl: './about-me.component.html',
   styleUrls: ['./about-me.component.scss'],
 })
 export class AboutMeComponent {
-  _result: string[] = [''];
+  private _badgeHeight = '18.75rem';
+  private _badgeWidth = '12.5rem';
+  private _skillList = [
+    'C',
+    'C++',
+    'Java',
+    'JavaScript',
+    'Node.JS',
+    'Python',
+    'Typescript',
+    'Go',
+    'Ruby',
+    'Haskell',
+    'Prolog',
+  ];
+  _profileImage =
+    './assets/about-me/me.jpeg';
 
-  usernameForm: FormGroup = this.fb.group({
+  _result = [''];
+
+  usernameFormInApp: FormGroup = this.fb.group({
     words: [''],
     specialCharacters: [''],
   });
 
-  constructor(
-    private fb: FormBuilder,
-    private httpClient: HttpClient,
-  ) {}
+  constructor(private fb: FormBuilder, private httpClient: HttpClient) {}
 
-  get angularVersion() {
-    return VERSION;
+  // ngAfterContentInit(): void {
+  //   const targeto: HTMLDivElement = document.getElementById(
+  //     'creds'
+  //   ) as HTMLDivElement;
+  //   const scriptEl = document.createElement('script');
+  //   scriptEl.src = 'https://cdn.credly.com/assets/utilities/embed.js';
+  //   scriptEl.async = true;
+  //   scriptEl.type = 'text/javascript';
+  //   targeto.appendChild(scriptEl);
+  // }
+
+  onTabChange($event: MatTabChangeEvent) {
+    if ($event.tab.textLabel === 'Credentials') {
+      const scriptEl = document.createElement('script');
+      scriptEl.src = 'https://cdn.credly.com/assets/utilities/embed.js';
+      scriptEl.async = true;
+      scriptEl.type = 'text/javascript';
+      const targeto = document.querySelector(
+        '#credential-div'
+      ) as HTMLDivElement;
+      targeto.append(scriptEl);
+    }
+  }
+
+  get nasaApiKey() {
+    return environment.nasaAPIKey;
   }
 
   get results() {
@@ -34,11 +74,32 @@ export class AboutMeComponent {
     this._result = result;
   }
 
-  onSubmit(_token: any) {
+  public get badgeHeight() {
+    return this._badgeHeight;
+  }
+  public set badgeHeight(value) {
+    this._badgeHeight = value;
+  }
+  public get badgeWidth() {
+    return this._badgeWidth;
+  }
+  public set badgeWidth(value) {
+    this._badgeWidth = value;
+  }
+
+  public get skillList() {
+    return this._skillList;
+  }
+  public set skillList(value) {
+    this._skillList = value;
+  }
+
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  onSubmit(_token: unknown) {
     const body = {
-      words: (this.usernameForm.get('words')?.value as string).split(','),
+      words: (this.usernameFormInApp.get('words')?.value as string).split(','),
       specials: (
-        this.usernameForm.get('specialCharacters')?.value as string
+        this.usernameFormInApp.get('specialCharacters')?.value as string
       ).split(','),
     };
     console.log(body);
@@ -51,7 +112,7 @@ export class AboutMeComponent {
             'Access-Control-Allow-Origin': '*',
             'Content-Type': 'application/json',
           },
-        },
+        }
       )
       .subscribe((results) => {
         this.results = results;