+++ src/environments/environment.local.ts
@@ -1,12 +1,7 @@
 export const environment = {
-  appUrl: 'jlz-portfolio.web.app',
   appCheckDebug: '4ad6c24e-aab2-40b5-9cd2-f7fe4291db01',
-  secretService:
-    'https://secretservice2ndgen-wa22s5z3va-uc.a.run.app/api/v2/secrets',
-  secretServiceLocal:
-    'http://127.0.0.1:4001/jlz-portfolio/us-central1/secretService2ndGen/api/v2/secrets',
-  production: false,
-  local: true,
+  appUrl: 'jlz-portfolio.web.app',
+  deployable: false,
   firebaseConfig: {
     apiKey: 'AIzaSyDuG6ojn89_8rvp4pSVpcmBwvT6D1zunQY',
     authDomain: 'jlz-portfolio.firebaseapp.com',
@@ -14,9 +9,20 @@ export const environment = {
     storageBucket: 'jlz-portfolio.appspot.com',
     messagingSenderId: '518070660509',
     appId: '1:518070660509:web:3c14202d7467504157d374',
-    measurementId: 'G-YSKMGT15LP',
+    measurementId: 'G-CNT03XSD81',
   },
+  local: true,
+  nasaAPIKey: 'vnfZmI8NbIb44DmpbbSAvlMhFz7Ipnyc1tq6xNLb',
+  production: false,
   recaptchaSiteKey: '6LdV5L8jAAAAAK1GaBQkmOq37fdxVQszw5x_iIV3',
+  serviceOptions: {
+    useServiceWorker: false,
+    productionServices: false,
+    postService:
+      'http://127.0.0.1:5001/jlz-portfolio/us-central1/jzPortfolioApp/api/v3/posts',
+    secretService:
+      'http://127.0.0.1:5001/jlz-portfolio/us-central1/jzPortfolioApp/api/v3/secrets',
+  },
 };
 
 /*