+++ src/environments/environment.prod.ts
@@ -1,12 +1,7 @@
 export const environment = {
-  appUrl: 'jlz-portfolio.web.app',
   appCheckDebug: false,
-  secretService:
-    'https://secretservice2ndgen-wa22s5z3va-uc.a.run.app/api/v2/secrets',
-  secretServiceLocal:
-    'http://127.0.0.1:4001/jlz-portfolio/us-central1/secretService2ndGen/api/v2/secrets',
-  production: true,
-  local: false,
+  appUrl: 'jlz-portfolio.web.app',
+  deployable: true,
   firebaseConfig: {
     apiKey: 'AIzaSyDuG6ojn89_8rvp4pSVpcmBwvT6D1zunQY',
     authDomain: 'jlz-portfolio.firebaseapp.com',
@@ -16,5 +11,16 @@ export const environment = {
     appId: '1:518070660509:web:3c14202d7467504157d374',
     measurementId: 'G-YSKMGT15LP',
   },
+  local: false,
+  nasaAPIKey: 'vnfZmI8NbIb44DmpbbSAvlMhFz7Ipnyc1tq6xNLb',
+  production: true,
   recaptchaSiteKey: '6LdV5L8jAAAAAK1GaBQkmOq37fdxVQszw5x_iIV3',
+  serviceOptions: {
+    useServiceWorker: true,
+    productionServices: true,
+    secretService:
+      'https://us-central1-jlz-portfolio.cloudfunctions.net/jzPortfolioApp/api/v3/secrets',
+    postService:
+      'https://us-central1-jlz-portfolio.cloudfunctions.net/jzPortfolioApp/api/v3/posts',
+  },
 };