+++ src/models/cache.model.ts
@@ -1,6 +1,0 @@
-import { Observable } from 'rxjs';
-
-export interface MediaCacheModel {
-  index: number;
-  url: Observable<string | null>;
-}