+++ CHANGELOG.md
@@ -1,8 +1,6 @@
 ## <small>2.0.7 (2023-07-15)</small>
 
-* chore: updates check (#283) ([1027673](https://github.com/Zwiqler94/jz-portfolio/commit/1027673)), closes [#283](https://github.com/Zwiqler94/jz-portfolio/issues/283)
-
-
+- chore: updates check (#283) ([1027673](https://github.com/Zwiqler94/jz-portfolio/commit/1027673)), closes [#283](https://github.com/Zwiqler94/jz-portfolio/issues/283)
 
 ## <small>2.0.6 (2023-06-23)</small>